27/11: associations pragmata kai postman requests kai ligo watchlist controler klpa

+++ src/main/java/com/unipi/christossiap/crypto_wallet_thesisassignment/configuration/auth/jwt/JwtTokenProvider.java
@@ -23,7 +23,7 @@ public class JwtTokenProvider {
     private String jwtSecret;
 
     @Value("${app.jwtExpirationInMs}")
-    private int jwtExpirationInMs;
+    private long jwtExpirationInMs;
 
     public String generateToken(Authentication authentication) {
         UserDetails userDetails = (UserDetails) authentication.getPrincipal();
